requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.13
django-bootstrap-v5==1.0.11
django-filter==24.3
django-widget-tweaks==1.5.0
fontawesomefree==6.6.0
gunicorn==23.0.0
idna==3.7
packaging==24.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.8.2

